feat(chat): 重构聊天模块并添加工具调用功能
- 重构了聊天模块的代码结构，使其更加清晰和模块化
- 添加了对工具调用的支持，使AI可以调用外部工具来完成任务
- 更新了示例代码，展示了如何使用新的工具调用功能
- 引入了新的Content类型来处理消息内容，提高了代码的灵活性
