+++ Cargo.toml
@@ -11,13 +11,18 @@ keywords = ["ai", "machine-learning", "openai", "library"]
 [dependencies]
 serde_json = "1.0.94"
 derive_builder = "0.20.0"
-reqwest = { version = "0.12", default-features = false, features = ["json", "stream", "multipart"], optional = true }
+reqwest = { version = "0.12", default-features = false, features = [
+    "json",
+    "stream",
+    "multipart",
+], optional = true }
 serde = { version = "1.0.157", features = ["derive"] }
 reqwest-eventsource = "0.6"
 tokio = { version = "1.26.0", features = ["full"] }
 anyhow = "1.0.70"
 futures-util = "0.3.28"
 bytes = "1.4.0"
+tracing = "0.1.41"
 
 [dev-dependencies]
 dotenvy = "0.15.7"
