+++ examples/chat_cli.rs
@@ -1,6 +1,6 @@
 use dotenvy::dotenv;
 use openai::{
-    chat::{ChatCompletion, ChatCompletionMessage, ChatCompletionMessageRole},
+    chat::{ChatCompletion, ChatCompletionMessage, ChatCompletionMessageRole, Content},
     Credentials,
 };
 use std::io::{stdin, stdout, Write};
@@ -13,7 +13,7 @@ async fn main() {
 
     let mut messages = vec![ChatCompletionMessage {
         role: ChatCompletionMessageRole::System,
-        content: Some("You are a large language model built into a command line interface as an example of what the `openai` Rust library made by Valentine Briese can do.".to_string()),
+        content: Some(Content::new_str("You are a large language model built into a command line interface as an example of what the `openai` Rust library made by Valentine Briese can do.")),
         ..Default::default()
     }];
 
@@ -26,7 +26,7 @@ async fn main() {
         stdin().read_line(&mut user_message_content).unwrap();
         messages.push(ChatCompletionMessage {
             role: ChatCompletionMessageRole::User,
-            content: Some(user_message_content),
+            content: Some(Content::new_str(&user_message_content)),
             ..Default::default()
         });
 
@@ -40,7 +40,7 @@ async fn main() {
         println!(
             "{:#?}: {}",
             &returned_message.role,
-            &returned_message.content.clone().unwrap().trim()
+            &returned_message.content.clone().unwrap()
         );
 
         messages.push(returned_message);
