+++ src/chat.rs
@@ -1,162 +1,14 @@
 //! Given a chat conversation, the model will return a chat completion response.
+pub mod modules;
+pub mod requests;
+pub mod types;
+pub mod utils;
 
-use super::{
-    openai_get, openai_get_with_query, openai_post, ApiResponseOrError, Credentials,
-    RequestPagination, Usage,
-};
-use crate::openai_request_stream;
-use derive_builder::Builder;
-use futures_util::StreamExt;
-use reqwest::Method;
-use reqwest_eventsource::{CannotCloneRequestError, Event, EventSource};
+pub use modules::*;
+pub use requests::*;
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use std::collections::HashMap;
-use tokio::sync::mpsc::{channel, Receiver, Sender};
-
-/// A full chat completion.
-pub type ChatCompletion = ChatCompletionGeneric<ChatCompletionChoice>;
-
-/// A delta chat completion, which is streamed token by token.
-pub type ChatCompletionDelta = ChatCompletionGeneric<ChatCompletionChoiceDelta>;
-
-#[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct ChatCompletionGeneric<C> {
-    #[serde(default)]
-    pub id: String,
-    pub object: String,
-    pub created: u64,
-    pub model: String,
-    pub choices: Vec<C>,
-    pub usage: Option<Usage>,
-}
-
-#[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct ChatCompletionChoice {
-    pub index: u64,
-    pub finish_reason: String,
-    pub message: ChatCompletionMessage,
-}
-
-#[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct ChatCompletionChoiceDelta {
-    pub index: u64,
-    pub finish_reason: Option<String>,
-    pub delta: ChatCompletionMessageDelta,
-}
-
-fn is_none_or_empty_vec<T>(opt: &Option<Vec<T>>) -> bool {
-    opt.as_ref().map(|v| v.is_empty()).unwrap_or(true)
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone, Eq, PartialEq, Default)]
-pub struct ChatCompletionMessage {
-    /// The role of the author of this message.
-    pub role: ChatCompletionMessageRole,
-    /// The contents of the message
-    ///
-    /// This is always required for all messages, except for when ChatGPT calls
-    /// a function.
-    pub content: Option<String>,
-    /// The name of the user in a multi-user chat
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    /// The function that ChatGPT called. This should be "None" usually, and is returned by ChatGPT and not provided by the developer
-    ///
-    /// [API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-function_call)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub function_call: Option<ChatCompletionFunctionCall>,
-    /// Tool call that this message is responding to.
-    /// Required if the role is `Tool`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tool_call_id: Option<String>,
-    /// Tool calls that the assistant is requesting to invoke.
-    /// Can only be populated if the role is `Assistant`,
-    /// otherwise it should be empty.
-    #[serde(skip_serializing_if = "is_none_or_empty_vec")]
-    pub tool_calls: Option<Vec<ToolCall>>,
-}
-
-/// Same as ChatCompletionMessage, but received during a response stream.
-#[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct ChatCompletionMessageDelta {
-    /// The role of the author of this message.
-    pub role: Option<ChatCompletionMessageRole>,
-    /// The contents of the message
-    pub content: Option<String>,
-    /// The name of the user in a multi-user chat
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    /// The function that ChatGPT called
-    ///
-    /// [API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-function_call)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub function_call: Option<ChatCompletionFunctionCallDelta>,
-    /// Tool call that this message is responding to.
-    /// Required if the role is `Tool`.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tool_call_id: Option<String>,
-    /// Tool calls that the assistant is requesting to invoke.
-    /// Can only be populated if the role is `Assistant`,
-    /// otherwise it should be empty.
-    #[serde(skip_serializing_if = "is_none_or_empty_vec")]
-    pub tool_calls: Option<Vec<ToolCall>>,
-}
-
-#[derive(Deserialize, Serialize, Clone, Debug, Eq, PartialEq)]
-pub struct ToolCall {
-    /// The ID of the tool call.
-    pub id: String,
-    /// The type of the tool. Currently, only `function` is supported.
-    pub r#type: String,
-    /// The function that the model called.
-    pub function: ToolCallFunction,
-}
-
-#[derive(Deserialize, Serialize, Clone, Debug, Eq, PartialEq)]
-pub struct ToolCallFunction {
-    /// The name of the function to call.
-    pub name: String,
-    /// The arguments to call the function with, as generated by the model in
-    /// JSON format.
-    /// Note that the model does not always generate valid JSON, and may
-    /// hallucinate parameters not defined by your function schema.
-    /// Validate the arguments in your code before calling your function.
-    pub arguments: String,
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone, Eq, PartialEq)]
-pub struct ChatCompletionFunctionDefinition {
-    /// The name of the function
-    pub name: String,
-    /// The description of the function
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    /// The parameters of the function formatted in JSON Schema
-    /// [API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-parameters)
-    /// [See more information about JSON Schema.](https://json-schema.org/understanding-json-schema/)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parameters: Option<Value>,
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
-pub struct ChatCompletionFunctionCall {
-    /// The name of the function ChatGPT called
-    pub name: String,
-    /// The arguments that ChatGPT called (formatted in JSON)
-    /// [API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-function_call)
-    pub arguments: String,
-}
-
-/// Same as ChatCompletionFunctionCall, but received during a response stream.
-#[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
-pub struct ChatCompletionFunctionCallDelta {
-    /// The name of the function ChatGPT called
-    pub name: Option<String>,
-    /// The arguments that ChatGPT called (formatted in JSON)
-    /// [API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-function_call)
-    pub arguments: Option<String>,
-}
+pub use types::*;
+pub use utils::*;
 
 #[derive(Deserialize, Serialize, Debug, Clone, Copy, Eq, PartialEq)]
 #[serde(rename_all = "lowercase")]
@@ -169,364 +21,12 @@ pub enum ChatCompletionMessageRole {
     Developer,
 }
 
-#[derive(Serialize, Builder, Debug, Clone)]
-#[builder(derive(Clone, Debug, PartialEq))]
-#[builder(pattern = "owned")]
-#[builder(name = "ChatCompletionBuilder")]
-#[builder(setter(strip_option, into))]
-pub struct ChatCompletionRequest {
-    /// ID of the model to use. Currently, only `gpt-3.5-turbo`, `gpt-3.5-turbo-0301` and `gpt-4`
-    /// are supported.
-    model: String,
-    /// The messages to generate chat completions for, in the [chat format](https://platform.openai.com/docs/guides/chat/introduction).
-    messages: Vec<ChatCompletionMessage>,
-    /// What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.
-    ///
-    /// We generally recommend altering this or `top_p` but not both.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    temperature: Option<f32>,
-    /// An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.
-    ///
-    /// We generally recommend altering this or `temperature` but not both.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    top_p: Option<f32>,
-    /// How many chat completion choices to generate for each input message.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    n: Option<u8>,
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    stream: Option<bool>,
-    /// Up to 4 sequences where the API will stop generating further tokens.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    stop: Vec<String>,
-    /// This feature is in Beta. If specified, our system will make a best effort to sample deterministically, such that repeated requests with the same seed and parameters should return the same result. Determinism is not guaranteed, and you should refer to the system_fingerprint response parameter to monitor changes in the backend.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    seed: Option<u64>,
-    /// The maximum number of tokens allowed for the generated answer. By default, the number of tokens the model can return will be (4096 - prompt tokens).
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    max_tokens: Option<u64>,
-    /// The maximum number of tokens allowed for the generated answer.
-    /// For reasoning models such as o1 and o3-mini, this does not include reasoning tokens.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    max_completion_tokens: Option<u64>,
-    /// Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model's likelihood to talk about new topics.
-    ///
-    /// [See more information about frequency and presence penalties.](https://platform.openai.com/docs/api-reference/parameter-details)
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    presence_penalty: Option<f32>,
-    /// Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim.
-    ///
-    /// [See more information about frequency and presence penalties.](https://platform.openai.com/docs/api-reference/parameter-details)
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    frequency_penalty: Option<f32>,
-    /// Modify the likelihood of specified tokens appearing in the completion.
-    ///
-    /// Accepts a json object that maps tokens (specified by their token ID in the tokenizer) to an associated bias value from -100 to 100. Mathematically, the bias is added to the logits generated by the model prior to sampling. The exact effect will vary per model, but values between -1 and 1 should decrease or increase likelihood of selection; values like -100 or 100 should result in a ban or exclusive selection of the relevant token.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    logit_bias: Option<HashMap<String, f32>>,
-    /// A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse. [Learn more](https://platform.openai.com/docs/guides/safety-best-practices/end-user-ids).
-    #[builder(default)]
-    #[serde(skip_serializing_if = "String::is_empty")]
-    user: String,
-    /// Describe functions that ChatGPT can call
-    /// The latest models of ChatGPT support function calling, which allows you to define functions that can be called from the prompt.
-    /// For example, you can define a function called "get_weather" that returns the weather in a given city
-    ///
-    /// [Function calling API Reference](https://platform.openai.com/docs/api-reference/chat/create#chat/create-functions)
-    /// [See more information about function calling in ChatGPT.](https://platform.openai.com/docs/guides/gpt/function-calling)
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    functions: Vec<ChatCompletionFunctionDefinition>,
-    /// A string or object of the function to call
-    ///
-    /// Controls how the model responds to function calls
-    ///
-    /// - "none" means the model does not call a function, and responds to the end-user.
-    /// - "auto" means the model can pick between an end-user or calling a function.
-    /// - Specifying a particular function via {"name":\ "my_function"} forces the model to call that function.
-    ///
-    /// "none" is the default when no functions are present. "auto" is the default if functions are present.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    function_call: Option<Value>,
-    /// An object specifying the format that the model must output. Compatible with GPT-4 Turbo and all GPT-3.5 Turbo models newer than gpt-3.5-turbo-1106.
-    /// Setting to { "type": "json_object" } enables JSON mode, which guarantees the message the model generates is valid JSON.
-    /// Important: when using JSON mode, you must also instruct the model to produce JSON yourself via a system or user message. Without this, the model may generate an unending stream of whitespace until the generation reaches the token limit, resulting in a long-running and seemingly "stuck" request. Also note that the message content may be partially cut off if finish_reason="length", which indicates the generation exceeded max_tokens or the conversation exceeded the max context length.
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    response_format: Option<ChatCompletionResponseFormat>,
-    /// The credentials to use for this request.
-    #[serde(skip_serializing)]
-    #[builder(default)]
-    credentials: Option<Credentials>,
-    /// Parameters unique to the Venice API.
-    /// https://docs.venice.ai/api-reference/api-spec
-    #[builder(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    venice_parameters: Option<VeniceParameters>,
-    /// Whether to store the completion for use in distillation or evals.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[builder(default)]
-    pub store: Option<bool>,
-}
-
-#[derive(Serialize, Debug, Clone, Eq, PartialEq)]
-pub struct VeniceParameters {
-    pub include_venice_system_prompt: bool,
-}
-
-#[derive(Serialize, Debug, Clone, Eq, PartialEq)]
-pub struct ChatCompletionResponseFormat {
-    /// Must be one of text or json_object (defaults to text)
-    #[serde(rename = "type")]
-    typ: String,
-}
-
-impl ChatCompletionResponseFormat {
-    pub fn json_object() -> Self {
-        ChatCompletionResponseFormat {
-            typ: "json_object".to_string(),
-        }
-    }
-
-    pub fn text() -> Self {
-        ChatCompletionResponseFormat {
-            typ: "text".to_string(),
-        }
-    }
-}
-
-impl<C> ChatCompletionGeneric<C> {
-    pub fn builder(
-        model: &str,
-        messages: impl Into<Vec<ChatCompletionMessage>>,
-    ) -> ChatCompletionBuilder {
-        ChatCompletionBuilder::create_empty()
-            .model(model)
-            .messages(messages)
-    }
-}
-
-#[derive(Serialize, Builder, Debug, Clone, Default)]
-#[builder(derive(Clone, Debug, PartialEq))]
-#[builder(pattern = "owned")]
-#[builder(name = "ChatCompletionMessagesRequestBuilder")]
-#[builder(setter(strip_option, into))]
-pub struct ChatCompletionMessagesRequest {
-    #[serde(skip_serializing)]
-    pub completion_id: String,
-
-    #[builder(default)]
-    #[serde(skip_serializing)]
-    pub credentials: Option<Credentials>,
-
-    #[builder(default)]
-    #[serde(flatten)]
-    pub pagination: RequestPagination,
-}
-
-/// A list of messages for a chat completion.
-#[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct ChatCompletionMessages {
-    pub data: Vec<ChatCompletionMessage>,
-    pub object: String,
-    pub first_id: Option<String>,
-    pub last_id: Option<String>,
-    pub has_more: bool,
-}
-
-impl ChatCompletion {
-    pub async fn create(request: ChatCompletionRequest) -> ApiResponseOrError<Self> {
-        let credentials_opt = request.credentials.clone();
-        openai_post("chat/completions", &request, credentials_opt).await
-    }
-
-    /// Get a stored completion.
-    pub async fn get(id: &str, credentials: Credentials) -> ApiResponseOrError<Self> {
-        let route = format!("chat/completions/{}", id);
-        openai_get(route.as_str(), Some(credentials)).await
-    }
-}
-
-impl ChatCompletionDelta {
-    pub async fn create(
-        request: ChatCompletionRequest,
-    ) -> Result<Receiver<Self>, CannotCloneRequestError> {
-        let credentials_opt = request.credentials.clone();
-        let stream = openai_request_stream(
-            Method::POST,
-            "chat/completions",
-            |r| r.json(&request),
-            credentials_opt,
-        )
-        .await?;
-        let (tx, rx) = channel::<Self>(32);
-        tokio::spawn(forward_deserialized_chat_response_stream(stream, tx));
-        Ok(rx)
-    }
-
-    /// Merges the input delta completion into `self`.
-    pub fn merge(
-        &mut self,
-        other: ChatCompletionDelta,
-    ) -> Result<(), ChatCompletionDeltaMergeError> {
-        if other.id.ne(&self.id) {
-            return Err(ChatCompletionDeltaMergeError::DifferentCompletionIds);
-        }
-        for other_choice in other.choices.iter() {
-            for choice in self.choices.iter_mut() {
-                if choice.index != other_choice.index {
-                    continue;
-                }
-                choice.merge(other_choice)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-impl ChatCompletionChoiceDelta {
-    pub fn merge(
-        &mut self,
-        other: &ChatCompletionChoiceDelta,
-    ) -> Result<(), ChatCompletionDeltaMergeError> {
-        if self.index != other.index {
-            return Err(ChatCompletionDeltaMergeError::DifferentCompletionChoiceIndices);
-        }
-        if self.delta.role.is_none() {
-            if let Some(other_role) = other.delta.role {
-                // Set role to other_role.
-                self.delta.role = Some(other_role);
-            }
-        }
-        if self.delta.name.is_none() {
-            if let Some(other_name) = &other.delta.name {
-                // Set name to other_name.
-                self.delta.name = Some(other_name.clone());
-            }
-        }
-        // Merge contents.
-        match self.delta.content.as_mut() {
-            Some(content) => {
-                match &other.delta.content {
-                    Some(other_content) => {
-                        // Push other content into this one.
-                        content.push_str(other_content)
-                    }
-                    None => {}
-                }
-            }
-            None => {
-                match &other.delta.content {
-                    Some(other_content) => {
-                        // Set this content to other content.
-                        self.delta.content = Some(other_content.clone());
-                    }
-                    None => {}
-                }
-            }
-        };
-
-        // merge function calls
-        // function call names are concatenated
-        // arguments are merged by concatenating them
-        match self.delta.function_call.as_mut() {
-            Some(function_call) => {
-                match &other.delta.function_call {
-                    Some(other_function_call) => {
-                        // push the arguments string of the other function call into this one
-                        match (&mut function_call.arguments, &other_function_call.arguments) {
-                            (Some(function_call), Some(other_function_call)) => {
-                                function_call.push_str(&other_function_call);
-                            }
-                            (None, Some(other_function_call)) => {
-                                function_call.arguments = Some(other_function_call.clone());
-                            }
-                            _ => {}
-                        }
-                    }
-                    None => {}
-                }
-            }
-            None => {
-                match &other.delta.function_call {
-                    Some(other_function_call) => {
-                        // Set this content to other content.
-                        self.delta.function_call = Some(other_function_call.clone());
-                    }
-                    None => {}
-                }
-            }
-        };
-        Ok(())
-    }
-}
-
-impl From<ChatCompletionDelta> for ChatCompletion {
-    fn from(delta: ChatCompletionDelta) -> Self {
-        ChatCompletion {
-            id: delta.id,
-            object: delta.object,
-            created: delta.created,
-            model: delta.model,
-            usage: delta.usage,
-            choices: delta
-                .choices
-                .iter()
-                .map(|choice| ChatCompletionChoice {
-                    index: choice.index,
-                    finish_reason: clone_default_unwrapped_option_string(&choice.finish_reason),
-                    message: ChatCompletionMessage {
-                        role: choice
-                            .delta
-                            .role
-                            .unwrap_or_else(|| ChatCompletionMessageRole::System),
-                        content: choice.delta.content.clone(),
-                        name: choice.delta.name.clone(),
-                        function_call: choice.delta.function_call.clone().map(|f| f.into()),
-                        tool_call_id: None,
-                        tool_calls: Some(Vec::new()),
-                    },
-                })
-                .collect(),
-        }
-    }
-}
-
-impl From<ChatCompletionFunctionCallDelta> for ChatCompletionFunctionCall {
-    fn from(delta: ChatCompletionFunctionCallDelta) -> Self {
-        ChatCompletionFunctionCall {
-            name: delta.name.unwrap_or("".to_string()),
-            arguments: delta.arguments.unwrap_or_default(),
-        }
-    }
-}
-
-impl ChatCompletionMessages {
-    /// Create a builder for fetching messages for a stored completion.
-    pub fn builder(completion_id: String) -> ChatCompletionMessagesRequestBuilder {
-        ChatCompletionMessagesRequestBuilder::create_empty()
-            .completion_id(completion_id.to_string())
-    }
-
-    /// Fetch messages for a stored completion.
-    pub async fn fetch(
-        request: ChatCompletionMessagesRequest,
-    ) -> ApiResponseOrError<ChatCompletionMessages> {
-        let route = format!("chat/completions/{}/messages", request.completion_id);
-        let credentials = request.credentials.clone();
-        openai_get_with_query(route.as_str(), &request, credentials).await
-    }
+#[derive(Deserialize, Serialize, Debug, Clone, Eq, PartialEq)]
+#[serde(rename_all = "lowercase")]
+pub enum ToolChoiceMode {
+    None,
+    Auto,
+    Required,
 }
 
 #[derive(Debug)]
@@ -554,50 +54,6 @@ impl std::fmt::Display for ChatCompletionDeltaMergeError {
 
 impl std::error::Error for ChatCompletionDeltaMergeError {}
 
-async fn forward_deserialized_chat_response_stream(
-    mut stream: EventSource,
-    tx: Sender<ChatCompletionDelta>,
-) -> anyhow::Result<()> {
-    while let Some(event) = stream.next().await {
-        let event = event?;
-        match event {
-            Event::Message(event) => {
-                let completion = serde_json::from_str::<ChatCompletionDelta>(&event.data)?;
-                tx.send(completion).await?;
-            }
-            _ => {}
-        }
-    }
-    Ok(())
-}
-
-impl ChatCompletionBuilder {
-    pub async fn create(self) -> ApiResponseOrError<ChatCompletion> {
-        ChatCompletion::create(self.build().unwrap()).await
-    }
-
-    pub async fn create_stream(
-        mut self,
-    ) -> Result<Receiver<ChatCompletionDelta>, CannotCloneRequestError> {
-        self.stream = Some(Some(true));
-        ChatCompletionDelta::create(self.build().unwrap()).await
-    }
-}
-
-impl ChatCompletionMessagesRequestBuilder {
-    /// Fetch messages for the specified completion.
-    pub async fn fetch(self) -> ApiResponseOrError<ChatCompletionMessages> {
-        ChatCompletionMessages::fetch(self.build().unwrap()).await
-    }
-}
-
-fn clone_default_unwrapped_option_string(string: &Option<String>) -> String {
-    match string {
-        Some(value) => value.clone(),
-        None => "".to_string(),
-    }
-}
-
 impl Default for ChatCompletionMessageRole {
     fn default() -> Self {
         Self::User
@@ -607,8 +63,11 @@ impl Default for ChatCompletionMessageRole {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::{Credentials, RequestPagination};
     use dotenvy::dotenv;
+    use serde_json::Value;
     use std::time::Duration;
+    use tokio::sync::mpsc::Receiver;
     use tokio::time::sleep;
 
     #[tokio::test]
@@ -620,7 +79,7 @@ mod tests {
             "gpt-3.5-turbo",
             [ChatCompletionMessage {
                 role: ChatCompletionMessageRole::User,
-                content: Some("Hello!".to_string()),
+                content: Some(Content::new_str("Hello!")),
                 name: None,
                 function_call: None,
                 tool_call_id: None,
@@ -643,7 +102,7 @@ mod tests {
                 .content
                 .as_ref()
                 .unwrap(),
-            "Hello! How can I assist you today?"
+            &Content::new_str("Hello! How can I assist you today?")
         );
     }
 
@@ -658,10 +117,9 @@ mod tests {
             "gpt-3.5-turbo",
             [ChatCompletionMessage {
                 role: ChatCompletionMessageRole::User,
-                content: Some(
-                    "What type of seed does Mr. England sow in the song? Reply with 1 word."
-                        .to_string(),
-                ),
+                content: Some(Content::new_str(
+                    "What type of seed does Mr. England sow in the song? Reply with 1 word.",
+                )),
                 name: None,
                 function_call: None,
                 tool_call_id: None,
@@ -685,7 +143,7 @@ mod tests {
                 .content
                 .as_ref()
                 .unwrap(),
-            "Love"
+            &Content::new_str("Love")
         );
     }
 
@@ -698,7 +156,7 @@ mod tests {
             "gpt-3.5-turbo",
             [ChatCompletionMessage {
                 role: ChatCompletionMessageRole::User,
-                content: Some("Hello!".to_string()),
+                content: Some(Content::new_str("Hello!")),
                 name: None,
                 function_call: None,
                 tool_call_id: None,
@@ -722,7 +180,7 @@ mod tests {
                 .content
                 .as_ref()
                 .unwrap(),
-            "Hello! How can I assist you today?"
+            &Content::new_str("Hello! How can I assist you today?")
         );
     }
 
@@ -736,7 +194,7 @@ mod tests {
             [
                 ChatCompletionMessage {
                     role: ChatCompletionMessageRole::User,
-                    content: Some("What is the weather in Boston?".to_string()),
+                    content: Some(Content::new_str("What is the weather in Boston?")),
                     name: None,
                     function_call: None,
                     tool_call_id: None,
@@ -797,51 +255,6 @@ mod tests {
         );
     }
 
-    #[tokio::test]
-    async fn chat_response_format_json() {
-        dotenv().ok();
-        let credentials = Credentials::from_env();
-        let chat_completion = ChatCompletion::builder(
-            "gpt-3.5-turbo",
-            [ChatCompletionMessage {
-                role: ChatCompletionMessageRole::User,
-                content: Some("Write an example JSON for a JWT header using RS256".to_string()),
-                name: None,
-                function_call: None,
-                tool_call_id: None,
-                tool_calls: Some(Vec::new()),
-            }],
-        )
-        .temperature(0.0)
-        .seed(1337u64)
-        .response_format(ChatCompletionResponseFormat::json_object())
-        .credentials(credentials)
-        .create()
-        .await
-        .unwrap();
-        let response_string = chat_completion
-            .choices
-            .first()
-            .unwrap()
-            .message
-            .content
-            .as_ref()
-            .unwrap();
-        #[derive(Deserialize, Eq, PartialEq, Debug)]
-        struct Response {
-            alg: String,
-            typ: String,
-        }
-        let response = serde_json::from_str::<Response>(response_string).unwrap();
-        assert_eq!(
-            response,
-            Response {
-                alg: "RS256".to_owned(),
-                typ: "JWT".to_owned()
-            }
-        );
-    }
-
     #[test]
     fn builder_clone_and_eq() {
         let builder_a = ChatCompletion::builder("gpt-4", [])
@@ -882,14 +295,13 @@ mod tests {
             [
                 ChatCompletionMessage {
                     role: ChatCompletionMessageRole::User,
-                    content: Some(
+                    content: Some(Content::new_str(
                         "What's 0.9102847*28456? \
                         reply in plain text, \
                         round the number to to 2 decimals \
                         and reply with the result number only, \
-                        with no full stop at the end"
-                            .to_string(),
-                    ),
+                        with no full stop at the end",
+                    )),
                     name: None,
                     function_call: None,
                     tool_call_id: None,
@@ -897,7 +309,7 @@ mod tests {
                 },
                 ChatCompletionMessage {
                     role: ChatCompletionMessageRole::Assistant,
-                    content: Some("Let me calculate that for you.".to_string()),
+                    content: Some(Content::new_str("Let me calculate that for you.")),
                     name: None,
                     function_call: None,
                     tool_call_id: None,
@@ -912,7 +324,7 @@ mod tests {
                 },
                 ChatCompletionMessage {
                     role: ChatCompletionMessageRole::Tool,
-                    content: Some("the result is 25903.061423199997".to_string()),
+                    content: Some(Content::new_str("the result is 25903.061423199997")),
                     name: None,
                     function_call: None,
                     tool_call_id: Some("the_tool_call".to_owned()),
@@ -937,7 +349,7 @@ mod tests {
                 .content
                 .as_ref()
                 .unwrap(),
-            "25903.06"
+            &Content::new_str("25903.06")
         );
     }
 
@@ -950,7 +362,7 @@ mod tests {
             "gpt-3.5-turbo",
             [ChatCompletionMessage {
                 role: ChatCompletionMessageRole::User,
-                content: Some("Hello!".to_string()),
+                content: Some(Content::new_str("Hello!")),
                 ..Default::default()
             }],
         )
@@ -988,7 +400,7 @@ mod tests {
 
         let user_message = ChatCompletionMessage {
             role: ChatCompletionMessageRole::User,
-            content: Some("Tell me a short joke".to_string()),
+            content: Some(Content::new_str("Tell me a short joke")),
             ..Default::default()
         };
 
@@ -1019,7 +431,7 @@ mod tests {
 
         let user_message = ChatCompletionMessage {
             role: ChatCompletionMessageRole::User,
-            content: Some("Tell me a short joke".to_string()),
+            content: Some(Content::new_str("Tell me a short joke")),
             ..Default::default()
         };
 
@@ -1067,42 +479,5 @@ mod tests {
         assert_eq!(retrieved_messages2.has_more, false);
         assert!(retrieved_messages2.first_id.is_none());
         assert!(retrieved_messages2.last_id.is_none());
-    }
-
-    #[tokio::test]
-    async fn chat_google() {
-        dotenv().ok();
-        let credentials = Credentials::from_env();
-
-        let chat_completion = ChatCompletion::builder(
-            "gemini-2.0-flash",
-            [ChatCompletionMessage {
-                role: ChatCompletionMessageRole::User,
-                content: Some("Hello!".to_string()),
-                name: None,
-                function_call: None,
-                tool_call_id: None,
-                tool_calls: Some(Vec::new()),
-            }],
-        )
-        .temperature(0.0)
-        .response_format(ChatCompletionResponseFormat::text())
-        .credentials(credentials)
-        .create()
-        .await
-        .unwrap();
-
-        assert_eq!(
-            chat_completion
-                .choices
-                .first()
-                .unwrap()
-                .message
-                .content
-                .as_ref()
-                .unwrap()
-                .trim(),
-            "Hello! How can I help you today?"
-        );
     }
 }
